chore: Extract config normalization
